mips_ejtag: Check if the ROCC bit can be set
Chips like the Ingenic JZ4730 (EJTAG v2.5) have a ROCC bit where writing
a 1 causes the bit to be set. This is contrary to the EJTAG v2.6
specification where is it said that a 1 is ignored by the hardware.

If we can set ROCC, we should not set it in ejtag_ctrl. We might miss
some resets, but mips_m4k_poll doesn't do much with that information
right now anyway.

Change-Id: Icbaac3555750757f09a4d55184df24f250b9723c
Signed-off-by: Daniel Glöckner <daniel-gl@gmx.net>
Signed-off-by: Jiri Kastner <cz172638@gmail.com>